modify cosmoctl to be testable
